Refactor AddItemCubit and AddItemState to improve structure and functionality
