+++ lib/user/super_admin/features/add_item/presentation/views/add_item_view.dart
@@ -1,7 +1,7 @@
 import 'package:final_project_iti/core/routing/route_export.dart';
 import 'package:final_project_iti/user/super_admin/features/add_item/presentation/widgets/add_item_app_bar.dart';
 import 'package:final_project_iti/user/super_admin/features/add_item/presentation/widgets/title_and_description_section.dart';
-import 'package:final_project_iti/user/super_admin/features/add_item/presentation/controller/track_cubit/add_item_cubit.dart';
+import 'package:final_project_iti/user/super_admin/features/add_item/presentation/controller/add_item_cubit/add_item_cubit.dart';
 
 import '../widgets/add_item_button.dart';
 import '../widgets/curriculum_item_list_view.dart';
